Commented printfs, suppresed warnings, changed dimensions to avoid segfaults on small screen sizes

+++ Tugas 1/framebuffer.c
@@ -1,4 +1,5 @@
 #include "framebuffer.h"
+#include <stdint.h>
 
 FrameBuffer initialize() {
     FrameBuffer fb;
@@ -26,7 +27,7 @@ FrameBuffer initialize() {
 
     fb.buffer = (char *)mmap(0, fb.screen_size, PROT_READ | PROT_WRITE,
                              MAP_SHARED, fb.fbfd, 0);
-    if ((int)fb.buffer == -1) {
+    if ((int)(uintptr_t)fb.buffer == -1) {
         perror("Error: failed to map framebuffer device to memory");
         exit(4);
     }
